tagged: fix lifetime on BerObjectHeader
The lifetime for BerObjectHeader<'hdr> was elided so it was not bound to
'a. However, it is clear that header should be bound to the input that
it is parsing.

This was generating a strange compiler warning for users as shown below.

error[E0495]: cannot infer an appropriate lifetime for lifetime
parameter 'hdr in function call due to conflicting requirements
...
   = note: expected `std::result::Result<(&[u8],
                        der_parser::ber::BerObject<'_>), _>`
              found `std::result::Result<(&[u8],
                        der_parser::ber::BerObject<'_>), _>`

Lifetime elision rules can be found here:
- https://doc.rust-lang.org/reference/lifetime-elision.html